fix: specify dtype=torch.float32 in torch.full to avoid type error in bernoulli
torch.bernoulli() requires float tensor but torch.full() defaults to Long.
Add explicit dtype=torch.float32 to fix RuntimeError.

+++ marl_framework/mapping/simulations.py
@@ -54,7 +54,7 @@ class Simulation:
     def get_noisy_map_section(sensor_noise, map_section, mode):
         accuracy = 1 - sensor_noise
         correctness = torch.bernoulli(
-                torch.full(np.shape(map_section), 1 - sensor_noise)
+                torch.full(np.shape(map_section), 1 - sensor_noise, dtype=torch.float32)
             )
         correctness = (
                 torch.reshape(correctness, np.shape(map_section)).cpu().numpy()
